giomodule: do not try to load modules from gio-module-dir in IOS
There is no gio-module-dir for ios. When it tries to load on a real
device it gets ignored because the path doesn't exist. When it tries
on the simulator, it will print assertions because the path might
exist in the user machine, but it will fail to load the modules